Public Interface

HDF5Vectors.HDF5Vectors — Module

This module implements an AbstractVector whose underlying data is stored in an HDF5 file.

It generally can store vectors of elements with fixed sizes, where that element is one of:

HDF5-compatible numeric type (signed/unsigned ints and floats)

Enum

SVector, SMatrix, and SArray of numeric types

Tuple of numeric types

bits-type consisting of any types on this list

general composite type consisting of any types on this list

Vector, Matrix, and Array of any types on this list, as long as their dimensions are always the same

String

Futher, it can serialize types to bytes or strings and store those in the HDF5 file. This allows it to store:

Any type that serializes

Vector, Matrix, and Array that are different lengths from element to element (via serialization)

It fulfills the general AbstractVector interface. Note, however, that iterating directly is slow; for far better speed, iterate on iterable(arr).

HDF5Vectors.AbstractHDF5Vector — Type

An abstract type intended as the parent for all type of HDF5 vectors. Subtypes should have a corresponding storage style and implement create_hdf5_vector, and at least these parts of the AbstractArray interface: length, setindex!, push!, getindex, and collect.

HDF5Vectors.AbstractHDF5VectorStorageStyle — Type

An abstract type intended as a parent for all HDF5 vector storage styles.

HDF5Vectors.CompositeStorageStyle — Type

Used to store composite structured information, like:

General tuple of types on this list

General named tuple of types on this list

General struct of types on this list

HDF5Vectors.ElementalStorageStyle — Type

Used to store "elemental" types – types that HDF5 can natively understand, including:

Int8, Int16, Int32, and Int64 (and unsigned forms)

Float32 and Float64

Enum

Char

Bits-type structs

String

HDF5Vectors.construct — Function

construct(type::Type, el)

TODO: Update. Given the el "element", as stored in the HDF5 file, the constructs the given type. What the element is depends on the storage style associated with the type.

HDF5Vectors.create_hdf5_vector — Method

create_hdf5_vector(group, name, el_type; kwargs...)

Creates the appropriate HDF5 vector type for the given element type, storing the vector in the given HDF5 groupin a new group/dataset,name`.

Optional keyword arguments:

dims: Tuple of the dimensions to use for a Vector, Matrix, or Array

chunk_length: Length of chunk to use (default 1000)

portable: True to maximize how "portable" the storage is (default true)

HDF5Vectors.create_hdf5_vector — Method

create_hdf5_vector(style, group, name, el_type; kwargs...)

Creates the appropriate HDF5 vector type for the given storage style and element type, storing the vector in the given HDF5 groupin a new group/dataset,name`.

Optional keyword arguments:

dims: Tuple of the dimensions to use for a Vector, Matrix, or Array

chunk_length: Length of chunk to use (default 1000)

portable: True to maximize how "portable" the storage is (default true)

HDF5Vectors.deconstruct — Function

deconstruct(type::Type, el)

TODO: Update.

Decomposes the given el element into the datatype used for storage in the HDF5 file. This depends on the storage style associated with the type of the element.

HDF5Vectors.iterable — Method

iterable(arr::AbstractHDF5Vector)

Returns an iterable type corresponding to the given HDF5 vector. This is generally much faster than iterating on the vector directly. That is, instead of [f(el) for el in arr], it is much faster to use [f(el) for el in iterable(arr)].

HDF5Vectors.load_hdf5_vector — Method

TODO

HDF5Vectors.load_hdf5_vector — Method

TODO

HDF5Vectors.storage_style — Method

storage_style(el_type::Type; kwargs...)

Returns the storage style intended for this type. Available styles include:

ElementalStorageStyle for scalars or non-portable bits-type structs

ArrayStorageStyle for arrays of known, consistent dimensions holding elemental types

CompositeStorageStyle for general structs

ByteArrayStorageStyle for arrays of inconsistent dimensions

JSONStorageStyle for serializing types to JSON strings

The default storage style for scalars and "non-portable" bits-type structs (more below) is ElementalStorageStyle. For vectors with known dimensions, ArrayStorageStyle is the default. For other structs (either non-bits-types or "portable"), the default is CompositeStorageStyle. For all other types, ByteArrayStorageStyle is the default storage style.

Array storage results in HDF5 files where the dataset has the dimensions of each element, plus one added dimension. For instance, if each element to be stored is an m-by-n array, then the HDF5 file will contain an m-by-n-by-p array, where element k is [:, :, k].

Structs can be stored in a "portable" way. For the a struct defined as:

struct MyType
    a::Int64
    b::Float64
end

the resulting HDF5 file would look like so:

/my_group/my_vector/arrays/a # a 1D array of Int64
/my_group/my_vector/arrays/b # a 1D array of Float64

This format is called "portable" because it is easy to interpret this dataset outside of Julia.

"Portability" is controlled by the portable keyword argument. When this is false, the above struct would be stored as:

/my_group/my_vector # a 1D array of custom type inferred from MyType

This uses the HDF5 type system via the HDF5.jl package to encode the type. The underlying data can still be interpreted outside of Julia, but it requires substantially more code to interpret the type information in a useful way. If you are only interested in loading the HDF5 in Julia, use portable = false, and the resulting storage will be faster. (Note that non-bits types cannot use the HDF5 type system and hence will always use the portable form.)

When the elements to be stored are themselves vectors, matrices, or arrays of known dimension, the user should provide those dimensions via the dims keyword argument. Otherwise, since the dimensions of an array are not known from its type, and it's not known if the user intends for dimensions to be consistent over time or not,

Keyword arguments:

portable: When true (the default), composite types like structs will be stored in a slower but more portable way. (For other types, this argument is ignored.)

dims: Sets the dimensions of Array types (otherwise, ignored), such as (3, 4) when each element is a 3-by-4 matrix.

Users can add a storage_style method for their custom types to allow them to express how their types out to be stored. E.g., if a type should always be serialized, then this would instruct Julia to use serialization to a byte array for the give type:

HDF5Vectors.storage_style(::Type{MyType}; kwargs...) = HDF5Vectors.ByteArrayStorageStyle()